+++ bika/lims/browser/worksheet/views/results.py
@@ -5,12 +5,10 @@
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
+from bika.lims import api
 from bika.lims import bikaMessageFactory as _
 from bika.lims.browser import BrowserView
-from bika.lims.browser.worksheet.tools import (checkUserAccess,
-                                               showRejectionMessage)
-from bika.lims.browser.worksheet.views import (AnalysesTransposedView,
-                                               AnalysesView)
+from bika.lims.browser.worksheet.tools import showRejectionMessage
 from bika.lims.config import WORKSHEET_LAYOUT_OPTIONS
 from bika.lims.utils import getUsers
 from plone.app.layout.globals.interfaces import IViewView
@@ -19,53 +17,103 @@ from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.utils import safe_unicode
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from zope.interface import implements
+from plone.memoize import view
+from bika.lims.permissions import EditWorksheet
+from bika.lims.permissions import ManageWorksheets
+from bika.lims.browser.worksheet.views import AnalysesTransposedView
 
 
 class ManageResultsView(BrowserView):
+    """Worksheet Manage Results View
+    """
     implements(IViewView)
     template = ViewPageTemplateFile("../templates/results.pt")
 
     def __init__(self, context, request):
-        BrowserView.__init__(self, context, request)
-        self.getAnalysts = getUsers(context, ['Manager', 'LabManager', 'Analyst'])
+        super(ManageResultsView, self).__init__(context, request)
+        self.icon = "{}/{}".format(
+            self.portal_url,
+            "/++resource++bika.lims.images/worksheet_big.png"
+        )
+
         self.layout_displaylist = WORKSHEET_LAYOUT_OPTIONS
 
     def __call__(self):
-
-        # Deny access to foreign analysts
-        if checkUserAccess(self.context, self.request) is False:
-            return []
-
+        # TODO: Refactor Worfklow
+        grant = self.is_edit_allowed() and self.is_manage_allowed()
+        if not grant:
+            redirect_url = api.get_url(self.context)
+            return self.request.response.redirect(redirect_url)
+        # TODO: Refactor this function call
         showRejectionMessage(self.context)
 
-        self.icon = self.portal_url + "/++resource++bika.lims.images/worksheet_big.png"
-
         # Save the results layout
-        rlayout = self.request.get('resultslayout', '')
+        rlayout = self.request.get("resultslayout", "")
         if rlayout and rlayout in WORKSHEET_LAYOUT_OPTIONS.keys() \
            and rlayout != self.context.getResultsLayout():
             self.context.setResultsLayout(rlayout)
             message = _("Changes saved.")
-            self.context.plone_utils.addPortalMessage(message, 'info')
+            self.context.plone_utils.addPortalMessage(message, "info")
 
-        # Here we create an instance of WorksheetAnalysesView
-        if self.context.getResultsLayout() == '2':
-            # Transposed view
-            self.Analyses = AnalysesTransposedView(self.context, self.request)
+        # Classic/Transposed View Switch
+        if self.context.getResultsLayout() == "1":
+            view = "analyses_classic_view"
+            self.Analyses = api.get_view(
+                view, context=self.context, request=self.request)
         else:
-            # Classic view
-            self.Analyses = AnalysesView(self.context, self.request)
+            # TODO: Refactor transposed view to new listing
+            self.Analyses = AnalysesTransposedView(self.context, self.request)
 
         self.analystname = self.context.getAnalystName()
-        self.instrumenttitle = self.context.getInstrument() and self.context.getInstrument().Title() or ''
+        self.instrumenttitle = self.get_instrument_title()
 
         # Check if the instruments used are valid
         self.checkInstrumentsValidity()
 
         return self.template()
 
+    def get_analysts(self):
+        """Returns Analysts
+        """
+        roles = ["Manager", "LabManager", "Analyst"]
+        return getUsers(self.context, roles)
+
+    @view.memoize
+    def get_instrument_title(self):
+        """Return the current instrument title
+        """
+        instrument = self.context.getInstrument()
+        if not instrument:
+            return ""
+        return api.get_title(instrument)
+
+    @view.memoize
+    def is_edit_allowed(self):
+        """Check if edit is allowed
+        """
+        checkPermission = self.context.portal_membership.checkPermission
+        return checkPermission(EditWorksheet, self.context)
+
+    @view.memoize
+    def is_manage_allowed(self):
+        """Check if manage is allowed
+        """
+        checkPermission = self.context.portal_membership.checkPermission
+        return checkPermission(ManageWorksheets, self.context)
+
+    @view.memoize
+    def is_assignment_allowed(self):
+        """Check if analyst assignment is allowed
+        """
+        if not self.is_manage_allowed():
+            return False
+        review_state = api.get_workflow_status_of(self.context)
+        edit_states = ["open", "attachment_due", "to_be_verified"]
+        return review_state in edit_states
+
     def getInstruments(self):
-        # TODO: Return only the allowed instruments for at least one contained analysis
+        # TODO: Return only the allowed instruments for at least one contained
+        # analysis
         bsc = getToolByName(self, 'bika_setup_catalog')
         items = [('', '')] + [(o.UID, o.Title) for o in
                               bsc(portal_type='Instrument',
@@ -76,36 +124,29 @@ class ManageResultsView(BrowserView):
         items.sort(lambda x, y: cmp(x[1].lower(), y[1].lower()))
         return DisplayList(list(items))
 
-    def isAssignmentAllowed(self):
-        workflow = getToolByName(self.context, 'portal_workflow')
-        review_state = workflow.getInfoFor(self.context, 'review_state', '')
-        edit_states = ['open', 'attachment_due', 'to_be_verified']
-        return review_state in edit_states \
-            and self.context.checkUserManage()
+    def get_wide_interims(self):
+        """Returns a dictionary with the analyses services from the current
+        worksheet which have at least one interim with 'Wide' attribute set to
+        true and that have not been yet submitted
 
-    def getWideInterims(self):
-        """ Returns a dictionary with the analyses services from the current
-            worksheet which have at least one interim with 'Wide' attribute
-            set to true and that have not been yet submitted
-            The structure of the returned dictionary is the following:
-            <Analysis_keyword>: {
-                'analysis': <Analysis_name>,
-                'keyword': <Analysis_keyword>,
-                'interims': {
-                    <Interim_keyword>: {
-                        'value': <Interim_default_value>,
-                        'keyword': <Interim_key>,
-                        'title': <Interim_title>
-                    }
+        The structure of the returned dictionary is the following:
+        <Analysis_keyword>: {
+            'analysis': <Analysis_name>,
+            'keyword': <Analysis_keyword>,
+            'interims': {
+                <Interim_keyword>: {
+                    'value': <Interim_default_value>,
+                    'keyword': <Interim_key>,
+                    'title': <Interim_title>
                 }
             }
+        }
         """
         outdict = {}
         allowed_states = ['assigned', 'unassigned']
         for analysis in self._getAnalyses():
-            wf = getToolByName(analysis, 'portal_workflow')
             # TODO Workflow - Analysis Use a query instead of this
-            if wf.getInfoFor(analysis, 'review_state') not in allowed_states:
+            if api.get_workflow_status_of(analysis) not in allowed_states:
                 continue
 
             if analysis.getKeyword() in outdict.keys():
@@ -115,43 +156,47 @@ class ManageResultsView(BrowserView):
             if not calculation:
                 continue
 
-            andict = {'analysis': analysis.Title(),
-                      'keyword': analysis.getKeyword(),
-                      'interims': {}}
+            andict = {
+                "analysis": analysis.Title(),
+                "keyword": analysis.getKeyword(),
+                "interims": {}
+            }
 
             # Analysis Service interim defaults
             for field in analysis.getInterimFields():
-                if field.get('wide', False):
-                    andict['interims'][field['keyword']] = field
+                if field.get("wide", False):
+                    andict["interims"][field["keyword"]] = field
 
             # Interims from calculation
             for field in calculation.getInterimFields():
-                if field['keyword'] not in andict['interims'].keys() \
-                   and field.get('wide', False):
-                    andict['interims'][field['keyword']] = field
+                if field["keyword"] not in andict["interims"].keys() \
+                   and field.get("wide", False):
+                    andict["interims"][field["keyword"]] = field
 
-            if andict['interims']:
+            if andict["interims"]:
                 outdict[analysis.getKeyword()] = andict
         return outdict
 
     def checkInstrumentsValidity(self):
-        """ Checks the validity of the instruments used in the Analyses
-            If an analysis with an invalid instrument (out-of-date or
-            with calibration tests failed) is found, a warn message
-            will be displayed.
+        """Checks the validity of the instruments used in the Analyses If an
+        analysis with an invalid instrument (out-of-date or with calibration
+        tests failed) is found, a warn message will be displayed.
         """
         invalid = []
         ans = self._getAnalyses()
         for an in ans:
             valid = an.isInstrumentValid()
             if not valid:
-                inv = '%s (%s)' % (safe_unicode(an.Title()), safe_unicode(an.getInstrument().Title()))
+                instrument = an.getInstrument()
+                inv = "%s (%s)" % (
+                    safe_unicode(an.Title()), safe_unicode(instrument.Title()))
                 if inv not in invalid:
                     invalid.append(inv)
         if len(invalid) > 0:
-            message = _("Some analyses use out-of-date or uncalibrated instruments. Results edition not allowed")
-            message = "%s: %s" % (message, (', '.join(invalid)))
-            self.context.plone_utils.addPortalMessage(message, 'warn')
+            message = _("Some analyses use out-of-date or uncalibrated "
+                        "instruments. Results edition not allowed")
+            message = "%s: %s" % (message, (", ".join(invalid)))
+            self.context.plone_utils.addPortalMessage(message, "warn")
 
     # TODO Department filtering
     def _getAnalyses(self):